feat(#51): total time float 형으로 변경

+++ finda-auth/src/main/kotlin/finda/findaauth/domain/student/model/Student.kt
@@ -8,5 +8,5 @@ data class Student(
     val classNum: Int,
     val num: Int,
     val name: String,
-    val totalVolunteerTime: Int = 0
+    val totalVolunteerTime: Float = 0f
 )
